sched/fair: Refactor packing eligible test.
The purpose of is_packing_eligible() test is to check if
we should throwaway the active cpu and use the idle CPU.
Refactor this function such a way that, the throw away
CPU (i.e active_cpu) is marked -1. The follow up patch
use this refactoring to add more checks in this function.

Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: I3affe2a6a114b14f82329c4809b8c2ffe6b94f3c