sched/fair: upadte adjust_cpus_for_packing().
Two changes:

- Prefer idle cpu over an active CPU if the idle CPU is not
in a deep c-state than WFI.

- if target_cpus_count > 1 means, we have found two active
CPUs. so we assume packing on this active cpu does not raise
the frequency any more. since we select the target_cpu as the
least loaded cpu, we can go ahead and check if it can accmmodate
the task without raising the frequency or not. so remove that
limitation as well.

Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
[render: minor fixups]
Signed-off-by: Zachariah Kennedy <zkennedy87@gmail.com>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: I78d4e1e61418e6da4862e0086010273cc78c42d8