sched: fair: balance for single core cluster.
Android will unset SD_LOAD_BALANCE for single core cluster domain and
for some product it is true to have a single core cluster and the MC
domain thus lacks the SD_LOAD_BALANCE flag. This will cause
select_task_rq_fair logic for that core. And the task will spin forever
in that core.

Bug: 141334320
Test: boot and see task on core7 scheduled correctly

Signed-off-by: Wei Wang <wvw@google.com>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: I397e27c92f1262b3ae019f264b3fe0fd461e047b